requirements.txt
fastapi
uvicorn[standard]
requests
python-jose[cryptography]
pydantic
pytest
